feature: 플레이어 수를 나타내는 클래스
PlayerCount
 - 최소 플레이어 수 검증 로직 분리
 - 객체 캐싱
 - Ladder가 Players를 갖고 있을 필요가 없음

+++ src/test/java/nextstep/ladder/domain/TestLineStrategy.java
@@ -8,10 +8,10 @@ import nextstep.ladder.domain.line.LineGenerateStrategy;
 
 public class TestLineStrategy {
     public static LineGenerateStrategy INVALID_ALL_LINE_STRATEGY = number -> Stream.generate(() -> Boolean.TRUE)
-            .limit(number)
+            .limit(number.toInt())
             .collect(Collectors.toList());
 
-    public static LineGenerateStrategy INVALID_TRUE_IN_SUCCESSION_STRATEGY = number -> IntStream.range(0, number)
+    public static LineGenerateStrategy INVALID_TRUE_IN_SUCCESSION_STRATEGY = number -> IntStream.range(0, number.toInt())
             .mapToObj(num -> num % 3 != 0)
             .collect(Collectors.toList());
 }
